+++ .gitignore
@@ -2,3 +2,4 @@ __pycache__/
 *.pkl
 *.index
 output.json
+training/all/*
