+++ train.py
@@ -40,7 +40,7 @@ for file in training_data:
 # Print statistics
 attempts = len(training_data)
 failures = len(failed_files)
-print("Number of files attempted to be parsed = " + str(attempts) + ".")
+print("Number of files attempted to be parsed = " + str(attempts))
 print("Number of failed files = " + str(failures) +
       ". Failure rate = " + "{:.2f}".format(failures/attempts*100) + "%")
 if failures:
@@ -50,16 +50,27 @@ if failures:
 
 encoding = tiktoken.get_encoding("cl100k_base")
 encoding = tiktoken.encoding_for_model("gpt-3.5-turbo")
-total_tokens = 0
-num_chunks = len(chunks)
+num_member_tokens = 0
+num_method_tokens = 0
+max_tokens = 0
+num_member_chunks = len(chunks)
+num_method_chunks = 0
 for chunk in chunks:
-    total_tokens = total_tokens + len(encoding.encode(str(chunk)))
-print("Number of chunks generated = " + str(num_chunks))
+    num_member_tokens = num_member_tokens + len(encoding.encode(str(chunk)))
+    if chunk['member'] == "method":
+        num_method_tokens = num_method_tokens + len(encoding.encode(str(chunk)))
+        num_method_chunks = num_method_chunks + 1
+print("Number of chunks generated = " + str(num_member_chunks))
 print("Average number of tokens per chunk = " + 
-      "{:.2f}".format(total_tokens/num_chunks))
+      "{:.2f}".format(num_member_tokens/num_member_chunks))
+print("Average number of tokens per method chunk = " +
+      "{:.2f}".format(num_method_tokens/num_method_chunks))
+print("Maximum token size = " + str(max_tokens))
 print("Chunks sample")
 print("-------------")
 for chunk in chunks[:10]:
     print(str(chunk))
 print("...")
+
+
